requirements.txt
streamlit == 1.23.0
pandas == 1.5.3
plotly == 5.13.0
numpy == 1.23.1
datetime == 5.0
altair == 4.2.2
pydeck == 0.8.0
google.cloud
google-auth-oauthlib